requirements.txt
asgiref==3.4.1
django-environ==0.4.5
django==3.2.5
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
